% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera-dims.R
\name{cameraDIMS}
\alias{cameraDIMS}
\title{Modified version of CAMERA so that it can be run on DI-MS dataset}
\usage{
cameraDIMS(data, params_iso, params_adduct, rule_type = "extended",
  rule_pth = NULL, rule_sep = "\\t", rule_export = FALSE)
}
\arguments{
\item{data}{data.frame = two column data frame consisting of mz values and intensity}

\item{params_iso}{list = a list of all the parameters used for the isotope calculations}

\item{rule_type}{character = [primary, extended, user]}

\item{rule_pth}{character = path of the rules used for adducts. only required for rule type 'user'}

\item{rule_sep}{character = seperator for rulepath}

\item{rule_export}{boolean = If TRUE will export the ruleset used}

\item{params_iso}{list = a list of all the parameters used for the adduct calculations}
}
\description{
Perform the CAMERA annotation functions on a an DI-MS dataset consisting of a column of mz values and a
corresponding column of intensity values.

This was a quick hack as we wanted to get CAMERA working on DI-MS without having to make an official change to the software.

Full documentation for CAMERA can be found here
https://bioconductor.org/packages/release/bioc/html/CAMERA.html
}
